Increasing number of batches when single precision isn't large enough to store the batch size * N (for adjacency graph & CSR row in array.
Will eventually expose an argument through Python to allow the user to control the precision of the outputs (and thus the precision of the internal structures.). It's important to allow the user to set this explicitly because there's a huge tradeoff to be made between a larger number of much faster batches and a smaller number of much slower batches.

+++ cpp/src/dbscan/vertexdeg/algo.h
@@ -32,7 +32,7 @@ namespace Algo {
 /**
  * Calculates the vertex degree array and the epsilon neighborhood adjacency matrix for the batch.
  */
-template <typename value_t, typename index_t = long>
+template <typename value_t, typename index_t = int>
 void launcher(const ML::cumlHandle_impl &handle, Pack<value_t, index_t> data,
               index_t startVertexId, index_t batchSize, cudaStream_t stream) {
   data.resetArray(stream, batchSize + 1);
